use config file in next.config.js
