+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Feature: Add storybook to aid in developing and designing complicated states.
 - Fix: Properly parse server errors into `ServerError` for use throughout the application to show the user better error messages.
 - Fix: Deterministic `buildId` based on git commit SHA in order to allow running multiple instances of the application behind a load balancer.
+- Fix: Update bugsnag installation to work properly with new config file.
 
 ### v3.0.1
 - hotfix: Checkout the master branch with submodules when building production docker container
