+++ src/dataflow/getLibraryData.ts
@@ -8,7 +8,6 @@ import {
 import { OPDSFeed, OPDSShelfLink } from "opds-feed-parser";
 import ApplicationError, { PageNotFoundError, ServerError } from "errors";
 import { flattenSamlMethod } from "utils/auth";
-import { APP_CONFIG } from "config";
 
 /**
  * Returns a function to construct a registry catalog link, which leads to a
