+++ src/types/globals.d.ts
@@ -1,0 +1,8 @@
+import { AppConfig } from "interfaces";
+
+declare global {
+  interface Window {
+    dataLayer: Array<any> | undefined;
+  }
+  const APP_CONFIG: AppConfig;
+}
