--- src/types/window.d.ts
@@ -1,3 +1,0 @@
-interface Window {
-  dataLayer: Array<any> | undefined;
-}
